+++ asdf/tests/test_helpers.py
@@ -43,4 +43,4 @@ def test_conversion_error(tmpdir):
 
     with pytest.raises(AsdfConversionWarning):
         with pytest.warns(AsdfWarning, match="Unable to locate schema file"):
-            assert_roundtrip_tree(tree, tmpdir, extensions=FooExtension())
+            assert_roundtrip_tree(tree, tmpdir, extensions=[FooExtension()])